Move source files to appropriate folders
