+++ src/rts/player.cpp
